requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
constantly==23.10.4
contextlib2==21.6.0
cryptography==44.0.0
daphne==4.1.2
Django==5.1.5
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-unfold==0.46.0
frozenlist==1.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
injector==0.22.0
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyOpenSSL==25.0.0
python-environ==0.4.54
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
yarl==1.18.3
zope.interface==7.2
